Switch to `int64_t` to make m$s$ w1nd0ws happy
Rant: At this point I'm not sure that a true C++ compiler and stdlib exist. There is
m$s$ C++ thats a different language from GNU C++.

Look at all these plebians using windows...

+++ include/gz/transport/Discovery.hh
@@ -788,7 +788,7 @@ namespace ignition
         sockaddr_in clntAddr;
         socklen_t addrLen = sizeof(clntAddr);
 
-        ssize_t received = recvfrom(this->sockets.at(0),
+        int64_t received = recvfrom(this->sockets.at(0),
               reinterpret_cast<raw_type *>(rcvStr),
               this->kMaxRcvStr, 0,
               reinterpret_cast<sockaddr *>(&clntAddr),
